refactor: move DuplicateManager and JobSearchConfig to modules
