+++ scraper.py
@@ -25,6 +25,8 @@ import psycopg2.extras
 from psycopg2 import sql
 import pandas as pd
 from jobspy import scrape_jobs
+from duplicate_manager import DuplicateManager
+from scraper_config import JobSearchConfig
 
 import warnings
 # Suppress pandas SQLAlchemy warnings for psycopg2 connections
@@ -988,259 +990,6 @@ class JobDatabase:
             self.conn.close()
             logger.info("Database connection closed")
 
-
-class DuplicateManager:
-    """Class to handle duplicate job detection and management."""
-    
-    def __init__(self, db: JobDatabase):
-        """Initialize with database connection.
-        
-        Args:
-            db: JobDatabase instance
-        """
-        self.db = db
-        self.site_preference = ['linkedin', 'indeed', 'google']
-    
-    def identify_duplicates(self) -> Tuple[List[List[Dict]], List[str], List[str]]:
-        """Identify duplicate jobs and determine which ones to keep/delete.
-        
-        Returns:
-            Tuple of (duplicate_groups, ids_to_delete, ids_to_keep)
-        """
-        duplicate_groups = self.db.get_duplicate_groups()
-        ids_to_delete = []
-        ids_to_keep = []
-        
-        for group in duplicate_groups:
-            best_job = self._select_best_job(group)
-            ids_to_keep.append(best_job['id'])
-            
-            for job in group:
-                if job['id'] != best_job['id']:
-                    ids_to_delete.append(job['id'])
-        
-        return duplicate_groups, ids_to_delete, ids_to_keep
-    
-    def _select_best_job(self, jobs: List[Dict]) -> Dict:
-        """Select the best job from a group of duplicates using the ranking algorithm.
-        
-        Args:
-            jobs: List of job dictionaries
-            
-        Returns:
-            The best job dictionary
-        """
-        candidates = jobs.copy()
-        
-        # Step 1: Filter by description (keep jobs with descriptions)
-        jobs_with_desc = [job for job in candidates if job.get('description') and job['description'].strip()]
-        if jobs_with_desc and len(jobs_with_desc) < len(candidates):
-            candidates = jobs_with_desc
-        
-        if len(candidates) == 1:
-            return candidates[0]
-        
-        # Step 2: Filter by Colorado location preference
-        colorado_jobs = [job for job in candidates 
-                        if job.get('location') and 
-                        (', CO' in job['location'] or 
-                         'Colorado' in job['location'] or
-                         ', co' in job['location'].lower() or
-                         'colorado' in job['location'].lower())]
-        if colorado_jobs and len(colorado_jobs) < len(candidates):
-            candidates = colorado_jobs
-        
-        if len(candidates) == 1:
-            return candidates[0]
-        
-        # Step 3: Filter by salary (keep jobs with min_amount > 0)
-        jobs_with_salary = [job for job in candidates 
-                           if job.get('min_amount') and job['min_amount'] > 0]
-        if jobs_with_salary and len(jobs_with_salary) < len(candidates):
-            candidates = jobs_with_salary
-        
-        if len(candidates) == 1:
-            return candidates[0]
-        
-        # Step 4: Select by highest min_amount
-        if len(candidates) > 1:
-            candidates_with_salary = [job for job in candidates 
-                                    if job.get('min_amount') and job['min_amount'] > 0]
-            if candidates_with_salary:
-                max_salary = max(job['min_amount'] for job in candidates_with_salary)
-                highest_salary_jobs = [job for job in candidates_with_salary 
-                                     if job['min_amount'] == max_salary]
-                if len(highest_salary_jobs) < len(candidates):
-                    candidates = highest_salary_jobs
-        
-        if len(candidates) == 1:
-            return candidates[0]
-        
-        # Step 5: Filter by remote status (if mixed, prefer remote=1)
-        remote_values = {job.get('is_remote', False) for job in candidates}
-        if False in remote_values and True in remote_values:
-            remote_jobs = [job for job in candidates if job.get('is_remote') == True]
-            if remote_jobs:
-                candidates = remote_jobs
-        
-        if len(candidates) == 1:
-            return candidates[0]
-        
-        # Step 6: Filter out United States search queries (prefer others)
-        non_us_jobs = [job for job in candidates 
-                      if not (job.get('search_query') and 
-                             'united states' in job['search_query'].lower())]
-        if non_us_jobs and len(non_us_jobs) < len(candidates):
-            candidates = non_us_jobs
-        
-        if len(candidates) == 1:
-            return candidates[0]
-        
-        # Step 7: Select by site preference
-        for preferred_site in self.site_preference:
-            site_jobs = [job for job in candidates if job.get('site', '').lower() == preferred_site]
-            if site_jobs:
-                candidates = site_jobs
-                break
-        
-        if len(candidates) == 1:
-            return candidates[0]
-        
-        # Step 8: Select by most recent date_posted
-        if len(candidates) > 1:
-            # Filter jobs that have valid dates
-            jobs_with_dates = [job for job in candidates 
-                             if job.get('date_posted') and job['date_posted'].strip()]
-            if jobs_with_dates:
-                # Sort by date_posted (assuming ISO format or comparable string format)
-                try:
-                    jobs_with_dates.sort(key=lambda x: x['date_posted'], reverse=True)
-                    return jobs_with_dates[0]
-                except:
-                    # If date sorting fails, continue to fallback
-                    pass
-        
-        # Fallback: return first job if no other criteria distinguishes them
-        return candidates[0]
-    
-    def delete_duplicate_jobs_directly(self, ids_to_delete: List[str]) -> int:
-        """Delete duplicate jobs directly from database without creating files.
-        
-        Args:
-            ids_to_delete: List of job IDs to delete
-            
-        Returns:
-            Number of jobs deleted
-        """
-        if not ids_to_delete:
-            return 0
-            
-        try:
-            with self.db.conn.cursor() as cursor:
-                cursor.execute("DELETE FROM scraped_jobs WHERE id = ANY(%s)", (ids_to_delete,))
-                rows_deleted = cursor.rowcount
-                self.db.conn.commit()
-                
-            logger.info(f"Deleted {rows_deleted} duplicate jobs directly")
-            return rows_deleted
-        except Exception as e:
-            logger.error(f"Error deleting duplicate jobs directly: {str(e)}")
-            return 0
-    
-    def create_delete_ids_file(self, ids_to_delete: List[str], filename: str = None) -> None:
-        """Create/overwrite file with IDs to delete.
-        
-        Args:
-            ids_to_delete: List of job IDs to delete
-            filename: Name of the file to create/overwrite
-        """
-        if filename is None:
-            filename = os.path.join(SCRIPT_DIR, "configs", "filters", "delete_ids.txt")
-        try:
-            config_dir = os.path.dirname(filename)
-            if config_dir and not os.path.exists(config_dir):
-                os.makedirs(config_dir)
-            
-            with open(filename, 'w') as f:
-                for job_id in ids_to_delete:
-                    f.write(f"{job_id}\n")
-            
-            logger.info(f"Created {filename} with {len(ids_to_delete)} IDs")
-            print(f"Created {filename} with {len(ids_to_delete)} IDs to delete")
-        except Exception as e:
-            logger.error(f"Error creating delete IDs file: {str(e)}")
-            print(f"Error creating delete IDs file: {str(e)}")
-
-
-class JobSearchConfig:
-    """Class to handle job search configuration from file."""
-    
-    def __init__(self, config_path: str = None):
-        """Initialize with configuration file path.
-        
-        Args:
-            config_path: Path to the configuration file
-        """
-        if config_path is None:
-            config_path = os.path.join(SCRIPT_DIR, "configs", "search", "job_search_config.json")
-        self.config_path = config_path
-        self.config = self._load_config()
-        
-    def _load_config(self) -> Dict:
-        """Load configuration from file.
-        
-        Returns:
-            Dictionary containing configuration
-        """
-        config_dir = os.path.dirname(self.config_path)
-        if config_dir and not os.path.exists(config_dir):
-            os.makedirs(config_dir)
-            
-        if not os.path.exists(self.config_path):
-            logger.warning(f"Config file {self.config_path} not found. Creating default configuration.")
-            default_config = {
-                "jobs": [
-                    {
-                        "name": "Software Engineer",
-                        "enabled": True,
-                        "parameters": {
-                            "site_name": ["indeed", "linkedin", "glassdoor"],
-                            "search_term": "software engineer",
-                            "location": "San Francisco, CA",
-                            "results_wanted": 100,
-                            "hours_old": 72,
-                            "country_indeed": "USA"
-                        }
-                    }
-                ],
-                "global": {
-                    "description_format": "markdown",
-                    "enforce_annual_salary": True,
-                    "verbose": 1
-                }
-            }
-            with open(self.config_path, 'w') as f:
-                json.dump(default_config, f, indent=4)
-            return default_config
-            
-        with open(self.config_path, 'r') as f:
-            return json.load(f)
-    
-    def get_job_configs(self) -> List[Dict]:
-        """Get all enabled job configurations.
-        
-        Returns:
-            List of enabled job configurations
-        """
-        return [job for job in self.config.get("jobs", []) if job.get("enabled", True)]
-    
-    def get_global_params(self) -> Dict:
-        """Get global parameters that apply to all searches.
-        
-        Returns:
-            Dictionary of global parameters
-        """
-        return self.config.get("global", {})
 
 
 class JobScraper:
